Add files via upload
This project involves the use of Principal Component Analysis(PCA) and K-means for optimal hedge identification. PCA is used to reduce the number of features in the data whiles maintaining majority of the variability whiles k-means is used to group the stocks or derivatives by volatility. The clusters are then used to determine optimal weights for the various types of stocks in a given portfolio.